requirements.txt
numpy==2.2.0
pillow==11.3.0
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
transformers==4.52.3
datasets==2.18.0
bitsandbytes==0.47.0
accelerate==1.3.0
opencv-python==4.11.0.86
qwen_vl_utils==0.0.14
unsloth==2025.9.9




